requirements.txt
amqp==2.5.2
billiard==3.6.1.0
celery==4.4.0rc2
Django==2.1.1
djangorestframework==3.10.3
importlib-metadata==1.1.0
kombu==4.6.6
more-itertools==8.0.0
pkg-resources==0.0.0
psycopg2==2.7.3.2
pytz==2019.2
vine==1.3.0
zipp==0.6.0
